+++ app/controllers/application_controller.rb
@@ -2,16 +2,5 @@ class ApplicationController < ActionController::Base
   # Only allow modern browsers supporting webp images, web push, badges, import maps, CSS nesting, and CSS :has.
   allow_browser versions: :modern
 
-  helper_method :current_user
-
-  def current_user
-    # If session[:user_id] is nil, set it to nil, otherwise find the user by id.
-    @current_user ||= session[:user_id] && User.find_by(id: session[:user_id])
-  end
-
-  def authenticate_user!
-    if current_user.blank?
-      redirect_to new_session_path, flash: { notice: "Sign In first!" }
-    end
-  end
+  include Authentication
 end
