+++ app/views/home/index.html.erb
@@ -1,3 +1,5 @@
 <h1>Home#index</h1>
 <p>Find me in app/views/home/index.html.erb</p>
-<%= button_to "Sign out", session_path(current_user), method: :delete %>
+<% if user_signed_in? %>
+  <%= button_to "Sign out", session_path, method: :delete %>
+<% end %>
